task 1

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -10,9 +10,9 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	size_t i;
+	unsigned int i;
 
 	if (array != null && action != null)
-		for( i = 0; i < size; i++)
+		for (i = 0; i < size; i++)
 			action(array[i]);
 }
